productservice/pyproject.toml
[tool.poetry]
name = "vis_23"
version = "0.1.0"
description = ""
authors = ["linxside <git@linux-inside.de>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^3.2"
djangorestframework = "^3.14.0"
mysqlclient = "^2.1.1"
requests = "^2.29.0"
PyYAML = "^6.0"
uritemplate = "^4.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
